(deosn't compile) extract index binds to separate file
