+++ src/intermediate_bindings/mod.rs
@@ -7,6 +7,9 @@
 //!to make ecCodes usage safer and easier,
 //!but they are unsafe as they operate on raw `codes_handle`.  
 
+#[cfg(feature = "ec_index")]
+pub(crate) mod codes_index;
+
 use std::{
     ffi::{CStr, CString},
     ptr::{self, addr_of_mut},
@@ -18,8 +21,6 @@ type _SYS_IO_FILE = eccodes_sys::__sFILE;
 #[cfg(not(target_os = "macos"))]
 type _SYS_IO_FILE = eccodes_sys::_IO_FILE;
 
-#[cfg(feature = "ec_index")]
-use eccodes_sys::codes_index;
 use eccodes_sys::{
     codes_context, codes_handle, codes_keys_iterator, codes_nearest, CODES_NEAREST_SAME_DATA,
     CODES_NEAREST_SAME_GRID, CODES_TYPE_BYTES, CODES_TYPE_DOUBLE, CODES_TYPE_LABEL,
@@ -329,91 +330,6 @@ pub unsafe fn codes_handle_new_from_message_copy(message_buffer: &[u8]) -> *mut 
         message_buffer.as_ptr().cast::<libc::c_void>(),
         message_buffer.len(),
     )
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_index_read(filename: &str) -> Result<*mut codes_index, CodesError> {
-    let filename = CString::new(filename).unwrap();
-    let context: *mut codes_context = ptr::null_mut(); //default context
-    let mut error_code: i32 = 0;
-
-    let codes_index = eccodes_sys::codes_index_read(context, filename.as_ptr(), &mut error_code);
-    codes_index.drop_in_place();
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(codes_index)
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_index_delete(index: *mut codes_index) -> Result<(), CodesError> {
-    eccodes_sys::codes_index_delete(index);
-    Ok(())
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_index_select_long(
-    index: *mut codes_index,
-    key: &str,
-    value: i64,
-) -> Result<(), CodesError> {
-    let key = CString::new(key).unwrap();
-    let error_code = eccodes_sys::codes_index_select_long(index, key.as_ptr(), value);
-
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(())
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_index_select_double(
-    index: *mut codes_index,
-    key: &str,
-    value: f64,
-) -> Result<(), CodesError> {
-    let key = CString::new(key).unwrap();
-    let error_code = eccodes_sys::codes_index_select_double(index, key.as_ptr(), value);
-
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(())
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_index_select_string(
-    index: *mut codes_index,
-    key: &str,
-    value: &str,
-) -> Result<(), CodesError> {
-    let key = CString::new(key).unwrap();
-    let value = CString::new(value).unwrap();
-    let error_code = eccodes_sys::codes_index_select_string(index, key.as_ptr(), value.as_ptr());
-
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(())
-}
-
-#[cfg(feature = "ec_index")]
-pub unsafe fn codes_handle_new_from_index(
-    index: *mut codes_index,
-) -> Result<*mut codes_handle, CodesError> {
-    let mut error_code: i32 = 0;
-
-    let codes_handle = eccodes_sys::codes_handle_new_from_index(index, &mut error_code);
-
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(codes_handle)
 }
 
 pub unsafe fn codes_keys_iterator_new(
